+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudDataModels.h
@@ -24,6 +24,8 @@
 #import <FitCloudKit/FitCloudClassroomModeSettingsModel.h>
 #import <FitCloudKit/FitCloudEmotionSettingsModel.h>
 #import <FitCloudKit/FitCloudWorkoutSummaryDataModel.h>
+#import <FitCloudKit/FitCloudTodayFortuneDataModel.h>
+#import <FitCloudKit/FitCloudTalismanDataModel.h>
 
 /// Specific client defines
 #import <FitCloudKit/Huashengda.h>